backend/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Werkzeug==3.1.3

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Optional: Rate Limiting
limits==3.9.0

# Security
flask-talisman==1.1.0

# Utilities
click==8.1.7

# Database (adjust as needed)
# psycopg2-binary==2.9.6  # For PostgreSQL
# pymysql==1.0.2  # For MySQL

# Development and Debugging
# ipdb==0.13.9
# flake8==7.1.1

# Optional: Logging and Monitoring
# blinker==1.9.0